修改一下电机驱动的参数，电机测试成功

+++ SMARTCAR/project/code/motor.h
@@ -5,6 +5,6 @@
 
 void motor_init(void);
 
-void motor_set_duty(int8 motor_left, int8 motor_right);
+void motor_set_duty(int16 motor_left, int16 motor_right);
 
 #endif
